stash refactor of aws code
