+++ cli/app.go
@@ -4,6 +4,7 @@ import (
 	"github.com/stelligent/mu/common"
 	"github.com/urfave/cli"
 	"io/ioutil"
+	"github.com/stelligent/mu/provider/aws"
 )
 
 // NewApp creates a new CLI app
@@ -37,6 +38,12 @@ func NewApp() *cli.App {
 
 		// initialize context
 		err := context.InitializeContext(c.String("profile"), c.String("region"), c.Bool("dryrun"))
+		if err != nil {
+			return err
+		}
+
+		// TODO: support initializing context from other cloud providers?
+		err := aws.InitializeContext(context, c.String("profile"), c.String("region"), c.Bool("dryrun"))
 		if err != nil {
 			return err
 		}
