+++ common/cluster.go
@@ -1,15 +1,6 @@
 package common
 
-import (
-	"fmt"
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/ecr"
-	"github.com/aws/aws-sdk-go/service/ecr/ecriface"
-	"github.com/aws/aws-sdk-go/service/ecs"
-	"github.com/aws/aws-sdk-go/service/ecs/ecsiface"
-	"strings"
-)
+import "github.com/aws/aws-sdk-go/service/ecs"
 
 // ContainerInstance represents the ECS container instance
 type ContainerInstance *ecs.ContainerInstance
@@ -30,75 +21,3 @@ type ClusterManager interface {
 	RepositoryAuthenticator
 }
 
-type ecsClusterManager struct {
-	ecsAPI ecsiface.ECSAPI
-	ecrAPI ecriface.ECRAPI
-}
-
-func newClusterManager(sess *session.Session) (ClusterManager, error) {
-	log.Debug("Connecting to ECS service")
-	ecsAPI := ecs.New(sess)
-
-	log.Debug("Connecting to ECR service")
-	ecrAPI := ecr.New(sess)
-
-	return &ecsClusterManager{
-		ecsAPI: ecsAPI,
-		ecrAPI: ecrAPI,
-	}, nil
-}
-
-// ListInstances get the instances for a specific cluster
-func (ecsMgr *ecsClusterManager) ListInstances(clusterName string) ([]ContainerInstance, error) {
-	ecsAPI := ecsMgr.ecsAPI
-
-	params := &ecs.ListContainerInstancesInput{
-		Cluster: aws.String(clusterName),
-	}
-
-	log.Debugf("Searching for container instances for cluster named '%s'", clusterName)
-
-	var instanceIds []*string
-	err := ecsAPI.ListContainerInstancesPages(params, func(page *ecs.ListContainerInstancesOutput, lastPage bool) bool {
-		for _, instanceID := range page.ContainerInstanceArns {
-			instanceIds = append(instanceIds, instanceID)
-		}
-		return true
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	describeParams := &ecs.DescribeContainerInstancesInput{
-		Cluster:            aws.String(clusterName),
-		ContainerInstances: instanceIds,
-	}
-	describeOut, _ := ecsAPI.DescribeContainerInstances(describeParams)
-
-	instances := make([]ContainerInstance, len(describeOut.ContainerInstances))
-	for _, instance := range describeOut.ContainerInstances {
-		instances = append(instances, instance)
-	}
-	return instances, nil
-}
-
-func (ecsMgr *ecsClusterManager) AuthenticateRepository(repoURL string) (string, error) {
-	ecrAPI := ecsMgr.ecrAPI
-
-	params := &ecr.GetAuthorizationTokenInput{}
-
-	log.Debug("Authenticating to ECR repo")
-
-	resp, err := ecrAPI.GetAuthorizationToken(params)
-	if err != nil {
-		return Empty, err
-	}
-
-	for _, authData := range resp.AuthorizationData {
-		if strings.HasPrefix(fmt.Sprintf("https://%s", repoURL), aws.StringValue(authData.ProxyEndpoint)) {
-			return aws.StringValue(authData.AuthorizationToken), nil
-		}
-	}
-
-	return Empty, fmt.Errorf("unable to find token for repo url:%s", repoURL)
-}
