+++ common/instance.go
@@ -1,10 +1,7 @@
 package common
 
 import (
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/aws/aws-sdk-go/service/ec2/ec2iface"
 )
 
 // InstanceLister for getting instances
@@ -18,36 +15,4 @@ type Instance *ec2.Instance
 // InstanceManager composite of all instance capabilities
 type InstanceManager interface {
 	InstanceLister
-}
-
-type ec2InstanceManager struct {
-	ec2API ec2iface.EC2API
-}
-
-func newInstanceManager(sess *session.Session) (InstanceManager, error) {
-	log.Debug("Connecting to EC2 service")
-	ec2API := ec2.New(sess)
-
-	return &ec2InstanceManager{
-		ec2API: ec2API,
-	}, nil
-}
-
-// ListInstances get the instances for a specific cluster
-func (ec2Mgr *ec2InstanceManager) ListInstances(instanceIds ...string) ([]Instance, error) {
-	ec2InputParameters := &ec2.DescribeInstancesInput{
-		InstanceIds: aws.StringSlice(instanceIds),
-	}
-
-	var instances []Instance
-	ec2Mgr.ec2API.DescribeInstancesPages(ec2InputParameters, func(page *ec2.DescribeInstancesOutput, lastPage bool) bool {
-		for _, reservation := range page.Reservations {
-			for _, instance := range reservation.Instances {
-				instances = append(instances, instance)
-			}
-		}
-		return true
-	})
-
-	return instances, nil
 }
