+++ provider/aws/log.go
@@ -1,0 +1,5 @@
+package aws
+
+import "github.com/op/go-logging"
+
+var log = logging.MustGetLogger("aws")
